:truck: Move functions that are only called in a single file into that file
