+++ tests/conftest.py
@@ -390,41 +390,3 @@ def mock_dataset_non_normalized_times(
         mock_data_path / "anemoi" / "mock_dataset_non_normalized_times.zarr",
         mock_data_non_normalized_times,
     )
-
-
-def make_varying_sic_stream(
-    *,
-    dist_grid: np.ndarray,
-    timesteps: int,
-    base_radius: float,
-    rng: np.random.Generator,
-    ring_width: float = 6.0,
-    noise_std: float = 0.03,
-    radius_oscillation_amplitude: float = 0.5,
-    radius_oscillation_frequency: float = 0.7,
-) -> np.ndarray:
-    """Vectorized [T, H, W] sea-ice concentration with oscillating coastline radius.
-
-    - Coastline radius varies sinusoidally over time about base_radius
-    - Concentration decays exponentially with distance outside the coastline
-    - Adds small Gaussian noise and masks land (inside radius) to 0.0
-    """
-    height, width = dist_grid.shape
-
-    t_idx = np.arange(timesteps, dtype=float)
-    radius_t = base_radius + radius_oscillation_amplitude * np.sin(
-        radius_oscillation_frequency * t_idx
-    )  # [T]
-
-    dist_b = dist_grid[None, :, :]  # [1,H,W]
-    radius_b = radius_t[:, None, None]  # [T,1,1]
-
-    outside = np.maximum(0.0, dist_b - radius_b)  # [T,H,W]
-    falloff = np.exp(-(outside / max(1.0, float(ring_width))))
-
-    noise = rng.normal(0.0, noise_std, size=(timesteps, height, width))
-    sic = np.clip(falloff + noise, 0.0, 1.0)
-
-    # Land mask: inside coastline set to 0.0 (temporary until NaN land masking)
-    sic[dist_b < radius_b] = 0.0
-    return sic.astype(np.float32)
