+++ tests/visualisations/test_layout.py
@@ -6,6 +6,7 @@ from typing import Any
 import matplotlib.pyplot as plt
 import numpy as np
 import pytest
+from matplotlib.axes import Axes
 
 from icenet_mp.visualisations import DEFAULT_SIC_SPEC
 from icenet_mp.visualisations.layout import (
@@ -17,7 +18,29 @@ from icenet_mp.visualisations.layout import (
     set_suptitle_with_box,
 )
 
-from .test_helper_plot_layout import axis_rectangle, rectangles_overlap
+EPSILON: float = 1e-6
+RECTANGLE = tuple[float, float, float, float]  # (Left, Bottom, Right, Top)
+
+
+def axis_rectangle(ax: Axes) -> RECTANGLE:
+    """Return (left, bottom, right, top) in figure-normalised coords [0, 1]."""
+    bbox = ax.get_position()
+    return (bbox.x0, bbox.y0, bbox.x1, bbox.y1)
+
+
+def rectangles_overlap(
+    rect_a: RECTANGLE,
+    rect_b: RECTANGLE,
+    *,
+    epsilon: float = EPSILON,
+) -> bool:
+    """True if two axis-aligned rectangles overlap (with tolerance)."""
+    la, ba, ra, ta = rect_a
+    lb, bb, rb, tb = rect_b
+    separated_h = ra <= lb + epsilon or rb <= la + epsilon
+    separated_v = ta <= bb + epsilon or tb <= ba + epsilon
+    return not (separated_h or separated_v)
+
 
 # Silence Matplotlib animation warning in this test module
 pytestmark = pytest.mark.filterwarnings(
